Fixed bug with session not being updated because user is logged in but jave no active tab or window.

+++ apps/scripts/management/commands/clean_comet_session.py
@@ -1,7 +1,10 @@
 """
-Deletes CometSessions that share the same session_key only if it is
-not the last object that has that session_key and only if the object
-has a timestamp that is older than the REQUEST_TIMEOUT.
+Deletes CometSessionIndexes (and related CometSessions) 
+if it was last updated more than 24 hours. Why 24 hours? 
+Remember that there is a stay signed in option that may conflict with
+the user clearing the browser data whilst still logged in, 
+which results in a dangling session_key. Also, if the user does not
+choose to stay signed in, then we still get a dangling session_key.
 
 These checks are in place so that currently logged in sessions are 
 not affected by this script.
@@ -11,27 +14,25 @@ from django.core.management.base import BaseCommand
 from django.utils import timezone
 
 from libs.dateutil.relativedelta import relativedelta
-from apps.comet.models import CometSession
-from repunch.settings import REQUEST_TIMEOUT
+from apps.comet.models import CometSession, CometSessionIndex
+
+LAST_UPDATED_THRESHOLD = 24 # in hours
 
 class Command(BaseCommand):
     def handle(self, *args, **options):\
-        # cache a copy of all the current comets
-        comets = CometSession.objects.all()
         now = timezone.now()
-        timedout_time = now + relativedelta(seconds=\
-            (REQUEST_TIMEOUT+10)*-1)
-        for comet in comets:
-            # first update the object manager
-            CometSession.objects.update()
-            
-            # 1. check if this is the last object with the session_key
-            # 2. only delete if this object has a timestamp that is
-            #    older than the REQUEST_TIMEOUT
-            if len(CometSession.objects.filter(session_key=\
-                comet.session_key) > 1 and\
-                comet.datetime < timedout_time:
-                comet.delete()
+        timedout_time = now + relativedelta(hours=\
+            -1*LAST_UPDATED_THRESHOLD)
+        # cache a copy of all the current comet indecies
+        cometis = CometSessionIndex.objects.all()
+        for cometi in cometis:
+            if cometi.last_updated < timedout_time:
+                cometi.delete()
+                # also delete associated cometsessions 
+                # (there shouldn't be any at all though)
+                for comet in CometSessionIndex.objects.filter(\
+                    session_key=cometi.session_key):
+                    comet.delete()
